requirements.txt
GitPython
python3-wget